refactor: rename sessiontunnel as tunnel
This code is now generic code to start a tunnel. It's not bound to
the session anymore so better to change its name and docs.
