+++ experiment/urlgetter/.gitignore
@@ -1,1 +1,1 @@
-/oonipsiphon
+/urlgetter-tunnel
